+++ java/src/main/java/com/xxAMIDOxx/xxSTACKSxx/Application.java
@@ -10,31 +10,8 @@ import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class Application {
-
-    private static final String PROJECT_URL = "https://github.com/amido/stacks-java";
-
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
-    }
-
-    /**
-     * OAS/Swagger Configuration
-     */
-    @Bean
-    public OpenAPI customOpenApi() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Menu API")
-                        .version("1.0")
-                        .description("APIs used to interact and manage menus for a restaurant")
-                        .contact(new Contact()
-                                .name("Amido")
-                                .url(PROJECT_URL)
-                                .email("stacks@amido.com"))
-                        .license(new License()
-                                .name("Apache 2.0")
-                                .url("http://springdoc.org")));
-
     }
 }
